improve `g.Go` (#2956)
