+++ contrib/config/kubecm/go.mod
@@ -3,7 +3,7 @@ module github.com/gogf/gf/contrib/config/kubecm/v2
 go 1.19
 
 require (
-	github.com/gogf/gf/v2 v2.5.3
+	github.com/gogf/gf/v2 v2.5.4
 	k8s.io/api v0.27.4
 	k8s.io/apimachinery v0.27.4
 	k8s.io/client-go v0.27.4
