+++ contrib/config/nacos/go.mod
@@ -3,7 +3,7 @@ module github.com/gogf/gf/contrib/config/nacos/v2
 go 1.18
 
 require (
-	github.com/gogf/gf/v2 v2.5.3
+	github.com/gogf/gf/v2 v2.5.4
 	github.com/nacos-group/nacos-sdk-go v1.1.4
 )
