+++ contrib/config/polaris/go.mod
@@ -3,8 +3,8 @@ module github.com/gogf/gf/contrib/config/polaris/v2
 go 1.18
 
 require (
-	github.com/gogf/gf/v2 v2.5.3
-	github.com/polarismesh/polaris-go v1.5.1
+	github.com/gogf/gf/v2 v2.5.4
+	github.com/polarismesh/polaris-go v1.5.3
 )
 
 require (
@@ -33,7 +33,7 @@ require (
 	github.com/natefinch/lumberjack v2.0.0+incompatible // indirect
 	github.com/olekukonko/tablewriter v0.0.5 // indirect
 	github.com/pkg/errors v0.9.1 // indirect
-	github.com/polarismesh/specification v1.3.2-alpha.2 // indirect
+	github.com/polarismesh/specification v1.4.0 // indirect
 	github.com/prometheus/client_golang v1.12.2 // indirect
 	github.com/prometheus/client_model v0.2.0 // indirect
 	github.com/prometheus/common v0.32.1 // indirect
