+++ contrib/drivers/pgsql/go.mod
@@ -3,7 +3,7 @@ module github.com/gogf/gf/contrib/drivers/pgsql/v2
 go 1.18
 
 require (
-	github.com/gogf/gf/v2 v2.5.3
+	github.com/gogf/gf/v2 v2.5.4
 	github.com/lib/pq v1.10.9
 )
