+++ contrib/registry/zookeeper/go.mod
@@ -4,7 +4,7 @@ go 1.18
 
 require (
 	github.com/go-zookeeper/zk v1.0.3
-	github.com/gogf/gf/v2 v2.5.3
+	github.com/gogf/gf/v2 v2.5.4
 	golang.org/x/sync v0.3.0
 )
