+++ contrib/rpc/grpcx/go.mod
@@ -3,8 +3,8 @@ module github.com/gogf/gf/contrib/rpc/grpcx/v2
 go 1.18
 
 require (
-	github.com/gogf/gf/contrib/registry/file/v2 v2.5.3
-	github.com/gogf/gf/v2 v2.5.3
+	github.com/gogf/gf/contrib/registry/file/v2 v2.5.4
+	github.com/gogf/gf/v2 v2.5.4
 	go.opentelemetry.io/otel v1.14.0
 	go.opentelemetry.io/otel/trace v1.14.0
 	google.golang.org/grpc v1.57.0
