+++ contrib/trace/jaeger/go.mod
@@ -3,7 +3,7 @@ module github.com/gogf/gf/contrib/trace/jaeger/v2
 go 1.18
 
 require (
-	github.com/gogf/gf/v2 v2.5.3
+	github.com/gogf/gf/v2 v2.5.4
 	go.opentelemetry.io/otel v1.14.0
 	go.opentelemetry.io/otel/exporters/jaeger v1.14.0
 	go.opentelemetry.io/otel/sdk v1.14.0
