+++ contrib/trace/otlphttp/go.mod
@@ -3,7 +3,7 @@ module github.com/gogf/gf/contrib/trace/otlphttp/v2
 go 1.20
 
 require (
-	github.com/gogf/gf/v2 v2.5.3
+	github.com/gogf/gf/v2 v2.5.4
 	go.opentelemetry.io/otel v1.16.0
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace v1.16.0
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp v1.16.0
