+++ util/gutil/gutil_goroutine.go
@@ -8,38 +8,22 @@ package gutil
 
 import (
 	"context"
-
-	"github.com/gogf/gf/v2/errors/gcode"
-	"github.com/gogf/gf/v2/errors/gerror"
 )
-
-// Func is the function which contains context parameter.
-type Func func(ctx context.Context)
-
-// RecoverFunc is the panic recover function which contains context parameter.
-type RecoverFunc func(ctx context.Context, exception error)
 
 // Go creates a new asynchronous goroutine function with specified recover function.
 //
 // The parameter `recoverFunc` is called when any panic during executing of `goroutineFunc`.
 // If `recoverFunc` is given nil, it ignores the panic from `goroutineFunc` and no panic will
 // throw to parent goroutine.
-func Go(ctx context.Context, goroutineFunc Func, recoverFunc RecoverFunc) {
+//
+// But, note that, if `recoverFunc` also throws panic, such panic will be thrown to parent goroutine.
+func Go(
+	ctx context.Context,
+	goroutineFunc func(ctx context.Context),
+	recoverFunc func(ctx context.Context, exception error),
+) {
 	if goroutineFunc == nil {
 		return
 	}
-	go func() {
-		defer func() {
-			if exception := recover(); exception != nil {
-				if recoverFunc != nil {
-					if v, ok := exception.(error); ok && gerror.HasStack(v) {
-						recoverFunc(ctx, v)
-					} else {
-						recoverFunc(ctx, gerror.NewCodef(gcode.CodeInternalPanic, "%+v", exception))
-					}
-				}
-			}
-		}()
-		goroutineFunc(ctx)
-	}()
+	go TryCatch(ctx, goroutineFunc, recoverFunc)
 }
